added basic functionality for task create, toggle, edit and delete

+++ sql/schema/5_tasks.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 ALTER TABLE tasks ADD COLUMN
-last_modified_at INTEGER NOT NULL DEFAULT(0);
+last_modified_at BIGINT NOT NULL DEFAULT(0);
 
 -- +goose Down
 ALTER TABLE TASKS
